+++ packages/router/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to `@gluonjs/router` are recorded here.
 
 ## [Unreleased]
 
+## [1.6.0] - 2026-07-28
+
+### Changed
+
+- Advanced the lockstep release for `@gluonjs/json-forms` while retaining this package's existing public contract.
+
 ## [1.5.0] - 2026-07-27
 
 ### Changed